Assert identical commit ordering between recorded and replay
This PR adds an assertion to `ProxyImpl::ScheduledActionCommit` verifying that the replay and recorded agree on the performed commits (and the source frames in which the commits happen).

This assertion should help us understand if the commits diverge (in case this assertion hits) or if the same commits are happening, but the commits contain different layers (old assertion keeps hitting).