Distinguish network responses from their redirects.
For #RUN-2343 (https://linear.app/replay/issue/RUN-2343/the-network-monitor-should-show-the-response-url)

Network requests that are redirected currently emit duplicate
network request events with the id of the original request.

Disitinguish between these by assigning them a different id.

However, take care to ensure that bookmarks, which are tracked by the
underlying "umbrella request id", are properly registered with each
network request in the redirect chain.