Add wait between layout and redraw during repaintGraphics.
For RUN-2643 (https://linear.app/replay/issue/RUN-2643)

This is a "better" fix for the original problem of stale
paints.

The patch here provides the trappings for later making this
timeout into an explicitly woken wait by the main thread,
but nothing invokes the wake-up, and we rely on the timeout
hitting and the layout finishing by then.